misc update

+++ ARDUINO_CODE/Lights.h
@@ -8,9 +8,9 @@ class LIGHT_CONTROLS {
             RIGHT = 0x40
         };
         LIGHT_CONTROLS(CANBUS_COMMUNICATOR *c);
-        void flash_lights(bool dipped, bool fog, uint8_t duration_msec);
-        void flash_indicator(INDICATOR i, uint8_t duration_msec);
-        void flash_hazard(uint8_t duration_msec);
+        void flash_lights(bool dipped, bool fog, int duration_msec);
+        void flash_indicator(INDICATOR i, int duration_msec);
+        void flash_hazard(int duration_msec);
     private:
         CANBUS_COMMUNICATOR *canbus;
         can_frame x;
